lab5 befor exer3

+++ test/sort.c
@@ -9,7 +9,7 @@
 
 #include "syscall.h"
 
-int A[1024];	/* size of physical memory; with code, we'll run out of space!*/
+short A[1024];	/* size of physical memory; with code, we'll run out of space!*/
 
 int
 main()
